Add dependency on the package to the plugin and handler directories
On first run (on Ubuntu 12.04 at least, Puppet 3.0.1) I hit an issue on
first run where the module tries to create directories in /etc/sensu
before they are created. /etc/sensu is not defined by the module but is
created by the sensu package. The addition of this dependency appears to
resolve the first run problem.