Add a notify between server and client when both are running
When running the client and the server on the same node, on first run
it's possible that the client starts before the server. This appears to
cause it not to appear until you manually (or otherwise) kick the
client. Adding this notification fixes that problem.

For a test case try https://github.com/garethr/sensu-playground. Without
this fix the first time the server instance is created the client
doesn't register. It's possible it's a bug in the (stripped down)
manifests too, if you spot that instead let me know.