Enable the installation and configuration of 3rd party dasboard
Currently in the module, the only dashboard available is sensu-dashboard.
Unfortunately, from sensu 0.13, sensu-dashboard isn't available anymore,
this leads to a run error when sensu::dashboard is set to true.

One can stick to $sensu::version = 0.12 but since sensu aims to distribute
3rd party dasbhoard in its repo, the module could benefit from it.

The following patchset aims to let one specify the dashboard system one wants
to set. So  $dashboard can be set to 'uchiwa', or any other dashboard name
available in sensu repos.