Use the command parameter when defined with source
Previously if command and source parameters were both specified, command was
in fact ignored. This is not helpful when the command might need additional
preamble or arguments to function.

Change the logic to use the value of the command parameter if it is defined,
(and therefore allow the operator to shoot themselves in the foot), otherwise
behave as before by setting the command to the concatenation of the install
path and the handler filename.