fix rabbitmq ssl config
The intent, per the comments in the module, is that $sensu::rabbitmq_ssl
would be set to true if you set either a certificate or a key. That var
defaults to false. Current code does not actually override this value.
When the sensu_rabbitmq_config is called, it directly references this var,
so unless you've specifically set it in your config, it's always false.
This makes it behave the way the module comments describe.