Fix handling of dependencies property.
If the dependencies propery is defined in the Puppet resource, the type
will alternate between adding and removing it on runs as under the
create method it's not added, while subsequently it's not included in
the list from check_args.

This causes the first run to not add it, the second (and each subsequent
even) run to add it in, and then each odd run to remove it as it's now
classed as a custom property instead and doesn't match that provided
by the resource.