Set subscribers to under instead of empty array.
The subscriptions property is an optional one and when set to [] it will
always give an empty array in the JSON exported. Setting to undef will
clean up any new configuration created (it shouldn't affect existing
configuration and puppet will consider undef/nil/empty arrays as
equivalent by the looks of it when :array_matching is set in the type).