Fix system_test (changes on second puppet run)
Missing caller the port= setter on sensu_client_config's provider's
create method. When the resource is first created (more precisely, when
the ensure property is out of sync) the only synched property is ensure,
which delegates to create on the provider. Because writing to the json
object of the provider that serves as the data for the json object is
only done in the provider setters, we must call them explicitly from
create.