Use new parent for boolean properties
Using Puppet::Property::Boolean does not work correctly. It munges the
value to (boolean) true and false. This causes an unfortunate problem.
Once a property is true, it's not possible to change it back to false.
This is because on ResourceHarness::sync_if_needed() the property value
is not even checked to be in sync of if the should value is false:

    if param.should && !param.safe_insync?(current_value)

This should probably read `if !param.should.nil?`. But since it doesn't,
this commit introduces a new superclass for the boolean properties that
normalizes to :true/:false instead. To avoid extra complexity in the
provider, wich has to work with actual booleans, the superclass also
overrides `set` and `retrieve` to shield the provider from having to do
converisons :true/:false => true/false and back.

This creates an additional problem, though. The should value retrieved
in the provider through `resource[:boolean_property]` will still be
:true/:false. Because the `create` methods in providers are littered
with statements such as

    self.<property name> = resource[:<property name>]

the setters in the providers will still be called with :true/:false.
This will be addressed in the next commit.