transform default client_port into an integer
By using the old default value which was a string and not an integer,
sensu-client would never start and you would got this message:

{"timestamp":"2015-12-06T21:18:52.299496+0000","level":"fatal","message":"client socket port must be an integer","object":{"name":"instack","address":"192.0.2.1","subscriptions":["undercloud_basic"],"socket":{"bind":"127.0.0.1","port":"3030"},"safe_mode":false,"keepalive":{}}}

The port value (defined by sensu::client_port here) must be an integer
and not a string.
This patch transforms the default value from a string to an integer so
we can run sensu-client out-of-the-box.