plugins should be hash and use create_resources
the plugins variable should be hash to allow multiple types of plugins to be installed and allow the install directory to be chose as this defaults to /etc/sensu which breaks windows support. this requires an addition of plugins_defaults to provide an easy way to provide defaults to override sensu::plugins variables such as install_path easily.