Disable user creation on osfamily = windows by default
Until better user management for windows installs can
be added this is to prevent attempting to create user/group
on windows platform regardless of the manage_user parameter.
Fixes issue sensu/sensu-puppet#617
* NOTE: This would have been better as a case statement in a
params.pp class but since that isn't implemented this is a
crude workaround for the time being. This is also the default requirement
for PR sensu/sensu-puppet#588