(GH-688) Default sensu-plugin gem to use sensu_gem provider
Without this patch the default behavior is to manage sensu-plugin using
the system `gem` command via the gem provider.  This is a problem
because this version combination is unknown and untested.  The embedded
ruby should be used for supportability and predictable behavior.

This patch changes the default behavior to use the embedded ruby version
at `/opt/sensu/embedded/bin/gem` via the `sensu_gem` provider shipped in
this module.

closes #688