(GH-646) Default sensu package version to installed
Without this patch the default package version for sensu is `latest`.
This is a problem for two reasons, first the package may be
automatically upgraded on Linux systems, which is a burden for
Operations when behavior changes.  Second, Windows does not have a
provider supporting the upgradable feature which results in an error in
the default case.

This path addresses both issues by defaulting the version to
`installed`.

For the Windows 2012r2 Vagrant box, the version is explicitly passed in
tests/sensu-client-windows.pp because
http://repositories.sensuapp.org/msi/ does not have sensu-latest.msi,
which is used in the default case.