(#638) Enable multiple Sensu Enterprise Dashboard API endpoints
Without this patch it is not possible to manage multiple
sensu::enterprise::dashboard::api resources in the same datacenter
because the resources conflict over the namevar.  This patchset
addresses the problem by changing the namevar to the `host` and
establishing the `datacenter` attribute.

An example of how to exercise this behavior is located in
tests/sensu-enterprise-dashboard-multi-api.pp

resolves #638
closes #651
closes #584