(maint) Manage Service[sensu-enterprise] with sensu-server-enterprise VM
Without this patch the sensu-server-enterprise Vagrant VM is not managing the
Service[sensu-enterprise] resource.  This is a problem for testing the behavior
changes made in #720.

This patch addresses the problem by changing the class parameters of the
provisioning manifest for sensu-server-enterprise.  The credentials for the
Sensu Enterprise repository are expected to be set in the shell environment
where `vagrant up` is executed.

For example, the following direnv `.envrc` is sufficient for this project:

```
export FACTER_SE_USER='1234567890'
export FACTER_SE_PASS='Password1'
```