(#658) Manage the when attribute of sensu filters
Without this patch sensu filters are unable to manage the [when
attribute](https://sensuapp.org/docs/0.26/reference/filters.html#when-attributes)

This patch implements the `when` property on the sensu_filter type and json
provider.  The `sensu::filter` defined type is also updated to support the
`when` parameter.

The behavior may be exercised with `vagrant up el7-client`.  This will produce
`/etc/sensu/conf.d/filters/offhours.json` with the content of:

    {
      "filters": {
        "offhours": {
          "attributes": {
            "client": {
              "environment": "production"
            }
          },
          "when": {
            "days": {
              "all": [
                {
                  "begin": "2:00 AM",
                  "end": "1:00 AM"
                }
              ]
            }
          },
          "negate": false
        }
      }
    }

Resolves #658