(#589) Add Chocolatey support for Windows
Without this patch package resource used to install sensu on Windows defaults
to the `msi` provider and can't easily be changed to the Chocolatey provider.
This patch addresses the problem by adding a new class parameter,
`sensu::windows_package_provider` which defaults to `undef`, the default
behavior of Puppet on windows, and can be set to `'chocolatey'`.

Closes #664
Resolves #589

Original Author: john.puskar <jpuskar@us.ibm.com>

The chocolatey package provider may be exercised with `vagrant up
win2012r2-client-chocolatey`.  There is no `senu` package in the public
chocolatey repository, however, so installation will fail with the following error:

    ==> win2012r2-client-chocolatey: Error: Execution of 'C:\ProgramData\chocolatey\bin\choco.exe install sensu -y ' returned 1: Ch
    ocolatey v0.10.7
    ==> win2012r2-client-chocolatey: Installing the following packages:
    ==> win2012r2-client-chocolatey: sensu
    ==> win2012r2-client-chocolatey: By installing you accept licenses for the packages.
    ==> win2012r2-client-chocolatey: sensu not installed. The package was not found with the source(s) listed.
    ==> win2012r2-client-chocolatey:  If you specified a particular version and are receiving this message, it is possible that the
     package name exists but the version does not.
    ==> win2012r2-client-chocolatey:  Version: ""
    ==> win2012r2-client-chocolatey:  Source(s): "https://chocolatey.org/api/v2/"
    ==> win2012r2-client-chocolatey: Chocolatey installed 0/1 packages. 1 packages failed.
    ==> win2012r2-client-chocolatey:  See the log for details (C:\ProgramData\chocolatey\logs\chocolatey.log).

Even with this error, the behavior of the chocolatey package provider is
exercised.