(#433) Reload Service[sensu_api] on RabbitMQ config changes
Without this patch the Sensu API service doesn't reload itself when RabbitMQ
configuration changes occur.  This patch addresses the problem by adding a
relationship between the rabbitmq configuration class and the service resource.

Note, both sensu::rabbitmq::config and sensu::redis::config are in the catalog
regardless of which is actually being configured.

Resolves #433