(#600) Reload Sensu API when check configurations change
Without this patch changes to sensu::check and sensu::check_config resources do
not automatically reload the Sensu API service.  This causes the checks to not
show up in the dashboard UI which is a problem for the end user.

This patch adds notification relationships between sensu_check and
sensu_check_config resources and Service[sensu-api].

Resolves #600