(#748) Add puppet-strings gem and dependencies
Build the documentation with `bundle exec rake puppet:strings`

The documentation is generated in every matrix cell for Travis.

RedCarpet is used to support Github Flavored Markdown.

The change to sensu_redis_config.rb is to avoid a bug in YARD, throwing
ArgumentError trying to parse the affected node in the syntax tree.  This patch
changes the code to be compatible with YARD.

Resolves #748