Add config property to sensu::check and sensu_check
Without this patch there is no way to specify arbitrary configuration at the top
level of the sensu check configuration object.  This patch addresses the problem
by adding a new properto to sensu::check and sensu_check named `config`.  This
property accepts a Hash map and merges all other configuration on top of this
base starting point.

This patch also adds a new mix-in method, `PuppetX::Sensu::SortHash#sort_hash`.
This method returns a hash sorted recursively, meaning all nested hash values
are sorted in the same manner.  This method is intended for use in the #flush
method of all Sensu providers.

See #764