(#783) Add sensu::check content parameter, use sensu::write_json
Without this patch arbitrary to-level configuration scope items cannot be
managed in a sensu check.  This patch addresses the problem by adding a
`content` parameter to the sensu::check defined type.

This patch changes the defined type to pass a JSON object to sensu::write_json
instead of passing each named property to the sensu_check custom type.
`sensu::write_json` is used to ensure all attributes in a configuration file are
manageable and arbitrary attributes need only be placed in the `content` hash
map.

The ultimate goal of this work is to eliminate the `custom` parameter and
eventually refactor all defined types in the Puppet module to use
sensu::write_json and this approach.

Resolves #783