Add dependencies for sensu_gem plugins
Previously, on upgrades to Sensu that installed a new embedded Ruby
installation, there was a chance that plugins installed with the `sensu_gem`
package provider would be evaluated BEFORE the Sensu package was upgraded
(which would mean that a second run of Puppet was necessary to get the plugins
installed into the new embedded Ruby installation). This commit adds a resource
chain ensuring that a change to the Sensu package will come before any package
resources using the `sensu_gem` provider (which would in turn notify the Sensu
services of their change).