Correct reference to services
Tests were failing because of this stanza (and specifically the last line):

```puppet
Package['sensu'] ~>
Package<| provider == 'sensu_gem' |> ~>
Service['sensu-client', 'sensu-server', 'sensu-api']
```

On cases where the api or client hasn't been installed there won't be any
resources for the sensu-api or sensu-client service. In its place is this line:

```puppet
~> Service<| title == 'sensu-client' or title == 'sensu-server' or title == 'sensu-api'|>
```

That collector will discover the resources if they're there and setup the
dependencies, or exit silently if they're not in the catalog. It's not the
prettiest, but it will do.